+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   Improved dark grey contrast.
 -   Ensure templated strings are localized in tooltips.
 -   Allow addition to be used as a unary operation.
+-   Don't tokenize negative numbers; treat them as a unary evaluation.
 
 ### Maintenance